+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 0.0.1
 
-- Initial Marketplace release of Relay.
+- Initial Marketplace release of Azure DevOps Relay.